+++ sockeye/arguments.py
@@ -739,6 +739,11 @@ def add_model_parameters(params):
 def add_training_args(params):
     train_params = params.add_argument_group("Training parameters")
 
+    train_params.add_argument('--decoder-only',
+                               action='store_true',
+                               help='Pre-train a decoder. This is currently for RNN decoders only. '
+                                    'Default: %(default)s.')
+
     train_params.add_argument('--batch-size', '-b',
                               type=int_greater_or_equal(1),
                               default=4096,