+++ CHANGELOG.md
@@ -10,7 +10,18 @@ Note that Sockeye has checks in place to not translate with an old model that wa
 
 Each version section may have have subsections for: _Added_, _Changed_, _Removed_, _Deprecated_, and _Fixed_.
 
+## [1.18.31]
+### Added
+- Pre-training the RNN decoder. Usage:
+  1. Train with flag `--decoder-only`.
+  2. Feed identical source/target training data.
+
+## [1.18.30]
+### Fixed
+- Preserving max output length for each sentence to allow having identical translations for both with and without batching. 
+
 ## [1.18.29]
+### Changed
 - No longer restrict the vocabulary to 50,000 words by default, but rather create the vocabulary from all words which occur at least `--word-min-count` times. Specifying `--num-words` explicitly will still lead to a restricted
   vocabulary.